refactor: add base redis compatible class and serializer
