+++ src/cachelib/serializers.py
@@ -55,29 +55,8 @@ class BaseSerializer:
             return data
 
 
-"""Default serializers for each cache type.
-
-The following classes can be used to further customize
-serialiation behaviour. Alternatively, any serializer can be
-overridden in order to use a custom serializer with a different
-strategy altogether.
-"""
-
-
-class UWSGISerializer(BaseSerializer):
-    """Default serializer for UWSGICache."""
-
-
-class SimpleSerializer(BaseSerializer):
-    """Default serializer for SimpleCache."""
-
-
-class FileSystemSerializer(BaseSerializer):
-    """Default serializer for FileSystemCache."""
-
-
-class RedisSerializer(BaseSerializer):
-    """Default serializer for RedisCache."""
+class BaseRedisSerializer(BaseSerializer):
+    """Base serializer for Redis compatible caches."""
 
     def dumps(self, value: _t.Any, protocol: int = pickle.HIGHEST_PROTOCOL) -> bytes:
         """Dumps an object into a string for redis, using pickle by default."""
@@ -101,29 +80,37 @@ class RedisSerializer(BaseSerializer):
             return value
 
 
-class ValkeySerializer(BaseSerializer):
+"""Default serializers for each cache type.
+
+The following classes can be used to further customize
+serialiation behaviour. Alternatively, any serializer can be
+overridden in order to use a custom serializer with a different
+strategy altogether.
+"""
+
+
+class UWSGISerializer(BaseSerializer):
+    """Default serializer for UWSGICache."""
+
+
+class SimpleSerializer(BaseSerializer):
+    """Default serializer for SimpleCache."""
+
+
+class FileSystemSerializer(BaseSerializer):
+    """Default serializer for FileSystemCache."""
+
+
+class RedisSerializer(BaseRedisSerializer):
+    """Default serializer for RedisCache."""
+
+    pass
+
+
+class ValkeySerializer(BaseRedisSerializer):
     """Default serializer for ValkeyCache."""
 
-    def dumps(self, value: _t.Any, protocol: int = pickle.HIGHEST_PROTOCOL) -> bytes:
-        """Dumps an object into a string for valkey, using pickle by default."""
-        return b"!" + pickle.dumps(value, protocol)
-
-    def loads(self, value: _t.Optional[bytes]) -> _t.Any:
-        """The reversal of :meth:`dump_object`. This might be called with
-        None.
-        """
-        if value is None:
-            return None
-        if value.startswith(b"!"):
-            try:
-                return pickle.loads(value[1:])
-            except pickle.PickleError:
-                return None
-        try:
-            return int(value)
-        except ValueError:
-            # before 0.8 we did not have serialization. Still support that.
-            return value
+    pass
 
 
 class DynamoDbSerializer(RedisSerializer):
