+++ src/cachelib/valkey.py
@@ -1,10 +1,10 @@
 import typing as _t
 
-from cachelib.base import BaseCache
+from cachelib.redis_base import BaseRedisCache
 from cachelib.serializers import ValkeySerializer
 
 
-class ValkeyCache(BaseCache):
+class ValkeyCache(BaseRedisCache):
     """Uses the Valkey key-value store as a cache backend.
 
     The first argument can be either a string denoting address of the Valkey
@@ -26,8 +26,6 @@ class ValkeyCache(BaseCache):
     Any additional keyword arguments will be passed to ``valkey.Valkey``.
     """
 
-    _read_client: _t.Any = None
-    _write_client: _t.Any = None
     serializer = ValkeySerializer()
 
     def __init__(
@@ -40,7 +38,6 @@ class ValkeyCache(BaseCache):
         key_prefix: _t.Optional[_t.Union[str, _t.Callable[[], str]]] = None,
         **kwargs: _t.Any,
     ):
-        BaseCache.__init__(self, default_timeout)
         if host is None:
             raise ValueError("ValkeyCache host parameter may not be None")
         if isinstance(host, str):
@@ -50,64 +47,12 @@ class ValkeyCache(BaseCache):
                 raise RuntimeError("no valkey module found") from err
             if kwargs.get("decode_responses", None):
                 raise ValueError("decode_responses is not supported by ValkeyCache.")
-            self._write_client = self._read_client = valkey.Valkey(
+            client = valkey.Valkey(
                 host=host, port=port, password=password, db=db, **kwargs
             )
         else:
-            self._read_client = self._write_client = host
-        self.key_prefix = key_prefix or ""
-
-    def _get_prefix(self) -> str:
-        return (
-            self.key_prefix if isinstance(self.key_prefix, str) else self.key_prefix()
-        )
-
-    def _normalize_timeout(self, timeout: _t.Optional[int]) -> int:
-        """Normalize timeout by setting it to default of 300 if
-        not defined (None) or -1 if explicitly set to zero.
-
-        :param timeout: timeout to normalize.
-        """
-        timeout = BaseCache._normalize_timeout(self, timeout)
-        if timeout == 0:
-            timeout = -1
-        return timeout
-
-    def get(self, key: str) -> _t.Any:
-        return self.serializer.loads(
-            self._read_client.get(f"{self._get_prefix()}{key}")
-        )
-
-    def get_many(self, *keys: str) -> _t.List[_t.Any]:
-        if self.key_prefix:
-            prefixed_keys = [f"{self._get_prefix()}{key}" for key in keys]
-        else:
-            prefixed_keys = list(keys)
-        return [self.serializer.loads(x) for x in self._read_client.mget(prefixed_keys)]
-
-    def set(self, key: str, value: _t.Any, timeout: _t.Optional[int] = None) -> _t.Any:
-        timeout = self._normalize_timeout(timeout)
-        dump = self.serializer.dumps(value)
-        if timeout == -1:
-            result = self._write_client.set(
-                name=f"{self._get_prefix()}{key}", value=dump
-            )
-        else:
-            result = self._write_client.setex(
-                name=f"{self._get_prefix()}{key}", value=dump, time=timeout
-            )
-        return result
-
-    def add(self, key: str, value: _t.Any, timeout: _t.Optional[int] = None) -> _t.Any:
-        timeout = self._normalize_timeout(timeout)
-        dump = self.serializer.dumps(value)
-        created = self._write_client.setnx(
-            name=f"{self._get_prefix()}{key}", value=dump
-        )
-        # handle case where timeout is explicitly set to zero
-        if created and timeout != -1:
-            self._write_client.expire(name=f"{self._get_prefix()}{key}", time=timeout)
-        return created
+            client = host
+        super().__init__(client, default_timeout, key_prefix)
 
     def set_many(
         self, mapping: _t.Dict[str, _t.Any], timeout: _t.Optional[int] = None
@@ -124,35 +69,3 @@ class ValkeyCache(BaseCache):
                 pipe.setex(name=f"{self._get_prefix()}{key}", value=dump, time=timeout)
         results = pipe.execute()
         return [k for k, was_set in zip(mapping.keys(), results) if was_set]
-
-    def delete(self, key: str) -> bool:
-        return bool(self._write_client.delete(f"{self._get_prefix()}{key}"))
-
-    def delete_many(self, *keys: str) -> _t.List[_t.Any]:
-        if not keys:
-            return []
-        if self.key_prefix:
-            prefixed_keys = [f"{self._get_prefix()}{key}" for key in keys]
-        else:
-            prefixed_keys = [k for k in keys]
-        self._write_client.delete(*prefixed_keys)
-        return [k for k in keys if not self.has(k)]
-
-    def has(self, key: str) -> bool:
-        return bool(self._read_client.exists(f"{self._get_prefix()}{key}"))
-
-    def clear(self) -> bool:
-        status = 0
-        if self.key_prefix:
-            keys = self._read_client.keys(self._get_prefix() + "*")
-            if keys:
-                status = self._write_client.delete(*keys)
-        else:
-            status = self._write_client.flushdb()
-        return bool(status)
-
-    def inc(self, key: str, delta: int = 1) -> _t.Any:
-        return self._write_client.incr(name=f"{self._get_prefix()}{key}", amount=delta)
-
-    def dec(self, key: str, delta: int = 1) -> _t.Any:
-        return self._write_client.incr(name=f"{self._get_prefix()}{key}", amount=-delta)
